+++ examples/3d/complex_one_color.rs
@@ -2,7 +2,7 @@ use bevy::prelude::*;
 use bevy_svg::prelude::*;
 
 fn main() {
-    App::build()
+    App::new()
         .insert_resource(Msaa { samples: 4 })
         .insert_resource(WindowDescriptor {
             title: "complex_one_color".to_string(),
@@ -12,7 +12,7 @@ fn main() {
         })
         .add_plugins(DefaultPlugins)
         .add_plugin(bevy_svg::prelude::SvgPlugin)
-        .add_startup_system(setup.system())
+        .add_startup_system(setup)
         .run();
 }
 
@@ -22,7 +22,7 @@ fn setup(
 ) {
     let svg = asset_server.load("asteroid_field.svg");
     commands.spawn_bundle(PerspectiveCameraBundle::new_3d());
-    commands.spawn_bundle(SvgBundle {
+    commands.spawn_bundle(Svg3dBundle {
         svg,
         origin: Origin::Center,
         transform: Transform {